tx/transmit.rs: colocate remember_prepared_recycle test with its production code
Phase 1 of Path A test redistribution (sibling PR to refactor/tests-rings).
Move `remember_prepared_recycle_tracks_only_shared_fill_recycles` from
tx/mod.rs::tests into a fresh `#[cfg(test)] mod tests` at the bottom
of tx/transmit.rs.

The new tests block also imports PROTO_TCP explicitly because
transmit.rs's production scope does not glob crate::afxdp::* (unlike
tx/mod.rs which does `use super::*;` at file head). PROTO_TCP is
private to crate::afxdp so descendants can import it directly.

Mechanical: 1 test, 42 lines moved. cargo test --bins 865/0/2.
