+++ userspace-dp/src/afxdp/tx/mod.rs
@@ -1051,48 +1051,6 @@ mod tests {
         assert_eq!(shared_recycles, vec![(7, 42)]);
     }
 
-    #[test]
-    fn remember_prepared_recycle_tracks_only_shared_fill_recycles() {
-        let mut in_flight_prepared_recycles = FastMap::default();
-
-        remember_prepared_recycle(
-            &mut in_flight_prepared_recycles,
-            &PreparedTxRequest {
-                offset: 41,
-                len: 64,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 0,
-                cos_queue_id: None,
-                dscp_rewrite: None,
-            },
-        );
-        remember_prepared_recycle(
-            &mut in_flight_prepared_recycles,
-            &PreparedTxRequest {
-                offset: 42,
-                len: 64,
-                recycle: PreparedTxRecycle::FillOnSlot(7),
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 0,
-                cos_queue_id: None,
-                dscp_rewrite: None,
-            },
-        );
-
-        assert_eq!(in_flight_prepared_recycles.len(), 1);
-        assert_eq!(
-            in_flight_prepared_recycles.get(&42),
-            Some(&PreparedTxRecycle::FillOnSlot(7))
-        );
-        assert!(!in_flight_prepared_recycles.contains_key(&41));
-    }
 
     #[test]
     fn clone_prepared_request_for_cos_returns_local_copy_with_metadata() {
